requirements.txt
selenium==4.13.0
Pillow==10.0.1
webdriver-manager==4.0.1
art==6.1